[FEAT] MemoResponse -> Keep 모델로 이동 및 메모 관련 API 추가
