+++ app/src/main/java/com/cocaine/myply/feature/data/datasource/remote/MyPlyService.kt
@@ -1,11 +1,7 @@
 package com.cocaine.myply.feature.data.datasource.remote
 
-import com.cocaine.myply.feature.data.model.SearchResponse
-import com.cocaine.myply.feature.data.model.SearchTagResponse
-import retrofit2.Retrofit
-import retrofit2.http.GET
-import retrofit2.http.Query
-import javax.inject.Inject
+import com.cocaine.myply.feature.data.model.*
+import retrofit2.http.*
 
 interface MyPlyService {
     @GET("tags/recommend")
@@ -13,4 +9,13 @@ interface MyPlyService {
 
     @GET("musics/search")
     suspend fun searchMusicPlayList(@Query("q") query: String, @Query("nextToken")nextPageToken: String? = null): SearchResponse
+
+    @GET("memos")
+    suspend fun getUserMemos(): UserMemoList
+
+    @PATCH("memos/{memoId}")
+    suspend fun updateMemo(@Path("memoId") memoId: String, @Body body: MemoUpdate): MemoResponse
+
+    @GET("memos/{memoId}")
+    suspend fun getMemo(@Path("memoId") memoId: String): MemoResponse
 }
