+++ app/src/main/java/com/cocaine/myply/feature/data/model/KeepModel.kt
@@ -1,0 +1,19 @@
+package com.cocaine.myply.feature.data.model
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+data class UserMemoList(val code: Int, val data: Memos?, val message: String?)
+
+data class Memos(val memos: List<MemoResponse>)
+
+@Parcelize
+data class MemoResponse(
+    val memoID: String,
+    val thumbnailURL: String,
+    val title: String,
+    val keywords: List<String>,
+    val body: String
+) : Parcelable
+
+data class MemoUpdate(val body: String)
