--- app/src/main/java/com/cocaine/myply/feature/data/model/MemoResponse.kt
@@ -1,9 +1,0 @@
-package com.cocaine.myply.feature.data.model
-
-data class MemoResponse(
-    val memoId: String,
-    val thumbnailURL: String,
-    val title: String,
-    val keywords: List<String>,
-    val body: String
-)
